+++ src/org/traccar/BasePipelineFactory.java
@@ -41,7 +41,7 @@ import org.traccar.events.AlertEventHandler;
 import org.traccar.helper.Log;
 import org.traccar.processing.ComputedAttributesHandler;
 import org.traccar.processing.CopyAttributesHandler;
-import org.traccar.processing.peripheralsensorprocessors.PeripheralSensorDataHandler;
+import org.traccar.processing.peripheralsensorprocessors.fuelsensorprocessors.FuelSensorDataHandler;
 
 import java.net.InetSocketAddress;
 
@@ -59,7 +59,7 @@ public abstract class BasePipelineFactory implements ChannelPipelineFactory {
     private HemisphereHandler hemisphereHandler;
     private CopyAttributesHandler copyAttributesHandler;
     private ComputedAttributesHandler computedAttributesHandler;
-    private PeripheralSensorDataHandler peripheralSensorDataHandler;
+    private FuelSensorDataHandler fuelSensorDataHandler;
 
     private CommandResultEventHandler commandResultEventHandler;
     private OverspeedEventHandler overspeedEventHandler;
@@ -171,7 +171,7 @@ public abstract class BasePipelineFactory implements ChannelPipelineFactory {
         }
 
         if (Context.getConfig().getBoolean("processing.peripheralSensorData.enable")) {
-            peripheralSensorDataHandler = new PeripheralSensorDataHandler();
+            fuelSensorDataHandler = new FuelSensorDataHandler();
         }
 
         if (Context.getConfig().getBoolean("event.enable")) {
@@ -239,8 +239,8 @@ public abstract class BasePipelineFactory implements ChannelPipelineFactory {
             pipeline.addLast("computedAttributes", computedAttributesHandler);
         }
 
-        if (peripheralSensorDataHandler != null) {
-            pipeline.addLast("peripheralSensorData", peripheralSensorDataHandler);
+        if (fuelSensorDataHandler != null) {
+            pipeline.addLast("peripheralSensorData", fuelSensorDataHandler);
         }
 
         if (Context.getDataManager() != null) {
